+++ packages/env-1password/CHANGELOG.md
@@ -2,6 +2,49 @@
 
 All notable changes to this project will be documented in this file. See [commit-and-tag-version](https://github.com/absolute-version/commit-and-tag-version) for commit guidelines.
 
+## [1.2.6](https://github.com/battis/qui-cli/compare/env-1password/1.2.5...env-1password/1.2.6) (2026-01-12)
+
+
+### Bug Fixes
+
+* include all peer dependencies ([4f386dd](https://github.com/battis/qui-cli/commit/4f386dd47861b91611c2fde49f55ee2955feaa05))
+
+## [1.2.5](https://github.com/battis/qui-cli/compare/env-1password/1.2.4...env-1password/1.2.5) (2026-01-02)
+
+
+### Bug Fixes
+
+* normalize peer dependencies ([cdd81b7](https://github.com/battis/qui-cli/commit/cdd81b7c4bea7c769021ff58177750c5a1369f76))
+
+## [1.2.4](https://github.com/battis/qui-cli/compare/env-1password/1.2.3...env-1password/1.2.4) (2025-12-26)
+
+
+### Bug Fixes
+
+* no longer logging service account token ([4303652](https://github.com/battis/qui-cli/commit/4303652a0ddbd6bf323dd548f07fd99bc95b2cc4))
+* update color-coding to current standards ([2fccf74](https://github.com/battis/qui-cli/commit/2fccf7433f729020022b7df85578e05576b47958))
+
+## [1.2.3](https://github.com/battis/qui-cli/compare/env-1password/1.2.2...env-1password/1.2.3) (2025-12-24)
+
+
+### Bug Fixes
+
+* handle inconsistencies of 1password-sdk sectionIds more gracefully ([bfa2be1](https://github.com/battis/qui-cli/commit/bfa2be1750da265676b31f7162f24e7d75a31502))
+
+## [1.2.2](https://github.com/battis/qui-cli/compare/env-1password/1.2.1...env-1password/1.2.2) (2025-12-24)
+
+
+### Bug Fixes
+
+* handle 1Password secret reference not in sections ([bd57bb6](https://github.com/battis/qui-cli/commit/bd57bb67f5027135425dfdd61fef5b3c1c2301c7))
+
+## [1.2.1](https://github.com/battis/qui-cli/compare/env-1password/1.2.0...env-1password/1.2.1) (2025-12-23)
+
+
+### Bug Fixes
+
+* fix crucial typo in documentation ([5fca554](https://github.com/battis/qui-cli/commit/5fca55433d9c8261122614c7786954bde779d63a))
+
 ## [1.2.0](https://github.com/battis/qui-cli/compare/env-1password/1.1.0...env-1password/1.2.0) (2025-12-23)